--- a/bench/micro/fiber_bench.ml
+++ b/bench/micro/fiber_bench.ml
@@ -84,3 +84,60 @@ let%bench "installing handlers" =
      in
      loop n)
   |> ignore
+
+let%bench_fun "Fiber.fork_and_join" =
+ fun () ->
+  Fiber.run
+    ~iter:(fun () -> assert false)
+    (let rec loop = function
+       | 0 -> Fiber.return ()
+       | n ->
+         let+ (), () =
+           Fiber.fork_and_join Fiber.return (fun () -> loop (n - 1))
+         in
+         ()
+     in
+     loop 1000)
+
+let%bench_fun "Fiber.fork_and_join_unit" =
+ fun () ->
+  Fiber.run
+    ~iter:(fun () -> assert false)
+    (let rec loop = function
+       | 0 -> Fiber.return ()
+       | n -> Fiber.fork_and_join_unit Fiber.return (fun () -> loop (n - 1))
+     in
+     loop 1000)
+
+let%bench_fun "Fiber.parallel_iter" =
+  let l = List.init 1000 ~f:Fun.id in
+  fun () ->
+    Fiber.run
+      ~iter:(fun () -> assert false)
+      (Fiber.parallel_iter l ~f:(fun _ -> Fiber.return ()))
+
+let%bench_fun "Fiber.parallel_map" =
+  let l = List.init 1000 ~f:Fun.id in
+  fun () ->
+    Fiber.run
+      ~iter:(fun () -> assert false)
+      (Fiber.parallel_map l ~f:Fiber.return)
+    |> ignore
+
+module M = Fiber.Make_map_traversals (Int.Map)
+
+let map =
+  List.init 1000 ~f:Fun.id
+  |> List.map ~f:(fun i -> (i, i))
+  |> Int.Map.of_list_exn
+
+let%bench "Fiber.Map.parallel_iter" =
+  Fiber.run
+    ~iter:(fun () -> assert false)
+    (M.parallel_iter map ~f:(fun _ _ -> Fiber.return ()))
+
+let%bench "Fiber.Map.parallel_map" =
+  Fiber.run
+    ~iter:(fun () -> assert false)
+    (M.parallel_map map ~f:(fun _ x -> Fiber.return x))
+  |> ignore