--- a/test/blackbox-tests/test-cases/coq/compose-two-scopes.t/run.t
+++ b/test/blackbox-tests/test-cases/coq/compose-two-scopes.t/run.t
@@ -3,6 +3,7 @@
   4 |  (theories a))
                  ^
   Error: Theory a not found
+  -> required by _build/default/b/b.v.d
   -> required by _build/default/b/b.vo
   -> required by _build/install/default/lib/coq/user-contrib/b/b.vo
   -> required by _build/default/cvendor.install