--- a/test/blackbox-tests/test-cases/coq/public-dep-on-private.t/run.t
+++ b/test/blackbox-tests/test-cases/coq/public-dep-on-private.t/run.t
@@ -4,6 +4,7 @@
                  ^^^^^^^
   Error: Theory "private" is private, it cannot be a dependency of a public
   theory. You need to associate "private" to a package.
+  -> required by _build/default/public/b.v.d
   -> required by _build/default/public/b.vo
   -> required by _build/install/default/lib/coq/user-contrib/public/b.vo
   -> required by _build/default/public.install