--- a/test/blackbox-tests/test-cases/enabled_if/eif-context_name.t/run.t
+++ b/test/blackbox-tests/test-cases/enabled_if/eif-context_name.t/run.t
@@ -34,6 +34,8 @@ dune >= 2.8
   18 |  (libraries bar))
                    ^^^
   Error: Library "bar" in _build/default is hidden (unsatisfied 'enabled_if').
+  -> required by _build/default/.bar_exe.eobjs/byte/dune__exe__Bar_exe.cmi
+  -> required by _build/default/.bar_exe.eobjs/native/dune__exe__Bar_exe.cmx
   -> required by _build/default/bar_exe.exe
   [1]