--- a/test/blackbox-tests/test-cases/enabled_if/eif-ocaml_version.t/run.t
+++ b/test/blackbox-tests/test-cases/enabled_if/eif-ocaml_version.t/run.t
@@ -10,5 +10,7 @@ This one is disabled (version too low)
                    ^^^^^^^^^^
   Error: Library "futurecaml" in _build/default is hidden (unsatisfied
   'enabled_if').
+  -> required by _build/default/.main2.eobjs/byte/dune__exe__Main2.cmi
+  -> required by _build/default/.main2.eobjs/native/dune__exe__Main2.cmx
   -> required by _build/default/main2.exe
   [1]