--- a/test/blackbox-tests/test-cases/libexec.t/run.t
+++ b/test/blackbox-tests/test-cases/libexec.t/run.t
@@ -315,7 +315,10 @@ But will fail when we release it, as it will need to run with -p:
   Error: Library "lib1" not found.
   -> required by %{libexec-private:lib1:lib1.ml} at lib2/dune:5
   -> required by _build/target/lib2/lib2.ml
-  -> required by _build/install/target/lib/public_lib2/lib2.ml
+  -> required by _build/target/lib2/.lib2.objs/byte/lib2.cmi
+  -> required by _build/target/lib2/.lib2.objs/native/lib2.cmx
+  -> required by _build/target/lib2/lib2.a
+  -> required by _build/install/target/lib/public_lib2/lib2.a
   -> required by _build/target/public_lib2.install
   -> required by alias install (context target)
   -> required by alias target (context target) in dune:5