--- a/test/blackbox-tests/test-cases/ocamldep-error-check.t
+++ b/test/blackbox-tests/test-cases/ocamldep-error-check.t
@@ -21,6 +21,7 @@ Dune uses ocamldep to prevent a module from depending on itself.
   of the library. Consequently, it should be the one depending on all the other
   modules in the library.
   -> required by _build/default/lib/.foo.objs/foo__Bar.impl.all-deps
+  -> required by _build/default/lib/.foo.objs/byte/foo__Bar.cmo
   -> required by _build/default/lib/foo.cma
   -> required by alias lib/all
   [1]