Using idea in #5197