@jeremiedimino 2021-11-25
Perfs are looking good:

main:
```
┌──────────────────────────────────────────────────┬──────────┬──────────┬───────────┬───────────┬────────────┐
│ Name                                             │ Time/Run │  mWd/Run │  mjWd/Run │  Prom/Run │ Percentage │
├──────────────────────────────────────────────────┼──────────┼──────────┼───────────┼───────────┼────────────┤
│ [bench/micro/fiber_bench.ml] bind                │   7.67us │  21.05kw │     4.54w │     4.54w │     11.65% │
│ [bench/micro/fiber_bench.ml] ivar                │  65.85us │  56.07kw │    16.37w │    16.37w │    100.00% │
│ [bench/micro/fiber_bench.ml] Var.set             │  41.66us │  32.06kw │ 1_269.22w │ 1_269.22w │     63.27% │
│ [bench/micro/fiber_bench.ml] Var.get             │  21.25us │  29.09kw │     7.90w │     7.90w │     32.27% │
│ [bench/micro/fiber_bench.ml] catching exceptions │  33.06us │  21.14kw │     9.76w │     9.76w │     50.21% │
│ [bench/micro/fiber_bench.ml] installing handlers │  51.56us │ 106.05kw │    38.59w │    38.59w │     78.30% │
└──────────────────────────────────────────────────┴──────────┴──────────┴───────────┴───────────┴────────────┘
```

this branch:

```
┌──────────────────────────────────────────────────┬──────────┬─────────┬───────────┬───────────┬────────────┐
│ Name                                             │ Time/Run │ mWd/Run │  mjWd/Run │  Prom/Run │ Percentage │
├──────────────────────────────────────────────────┼──────────┼─────────┼───────────┼───────────┼────────────┤
│ [bench/micro/fiber_bench.ml] bind                │   8.05us │ 21.07kw │     4.41w │     4.41w │     16.74% │
│ [bench/micro/fiber_bench.ml] ivar                │  35.65us │ 51.08kw │    16.03w │    16.03w │     74.19% │
│ [bench/micro/fiber_bench.ml] Var.set             │  48.06us │ 45.07kw │ 1_635.86w │ 1_635.86w │    100.00% │
│ [bench/micro/fiber_bench.ml] Var.get             │  27.38us │ 32.12kw │    10.37w │    10.37w │     56.98% │
│ [bench/micro/fiber_bench.ml] catching exceptions │  28.76us │ 22.13kw │   321.48w │   321.48w │     59.84% │
│ [bench/micro/fiber_bench.ml] installing handlers │  34.36us │ 77.07kw │    19.68w │    19.68w │     71.49% │
└──────────────────────────────────────────────────┴──────────┴─────────┴───────────┴───────────┴────────────┘
```

The difference for `bind` is noise: it's exactly the same implem between the two branches. But ivar got quite a bit faster. To be completely honest, I did some other optimisations on the representation. But it's nice that we can get similar to better perf with a more functional implementation.