@rgrinberg 2021-11-25
One thing omitted in the benchmark suite is the cost of fork. I doubt it's going to change much, but you never know.

How did get/set get slower?! It's not by much, but I would have expected it to be at least a little faster.