@jeremiedimino 2021-11-25
So for get/set it's expected as we are now creating this temporary `Get_var ...` or `Set_var ...` constructor. For fork, yeah that's a good point I need to try that.