@jeremiedimino 2021-11-25
I got the numbers down a little:

```
┌──────────────────────────────────────────────────┬──────────┬─────────┬───────────┬───────────┬────────────┐
│ Name                                             │ Time/Run │ mWd/Run │  mjWd/Run │  Prom/Run │ Percentage │
├──────────────────────────────────────────────────┼──────────┼─────────┼───────────┼───────────┼────────────┤
│ [bench/micro/fiber_bench.ml] bind                │   7.56us │ 21.07kw │     4.50w │     4.50w │     16.71% │
│ [bench/micro/fiber_bench.ml] ivar                │  33.27us │ 51.08kw │    16.87w │    16.87w │     73.50% │
│ [bench/micro/fiber_bench.ml] Var.set             │  45.27us │ 42.08kw │ 1_454.23w │ 1_454.23w │    100.00% │
│ [bench/micro/fiber_bench.ml] Var.get             │  23.89us │ 32.12kw │     9.59w │     9.59w │     52.77% │
│ [bench/micro/fiber_bench.ml] catching exceptions │  28.95us │ 22.13kw │   321.10w │   321.10w │     63.95% │
│ [bench/micro/fiber_bench.ml] installing handlers │  34.32us │ 77.07kw │    19.64w │    19.64w │     75.81% │
└──────────────────────────────────────────────────┴──────────┴─────────┴───────────┴───────────┴────────────┘
```