@jeremiedimino 2021-11-25
A tad more by tweaking the jobs representation:

```
┌──────────────────────────────────────────────────┬──────────┬─────────┬───────────┬───────────┬────────────┐
│ Name                                             │ Time/Run │ mWd/Run │  mjWd/Run │  Prom/Run │ Percentage │
├──────────────────────────────────────────────────┼──────────┼─────────┼───────────┼───────────┼────────────┤
│ [bench/micro/fiber_bench.ml] bind                │   7.50us │ 21.07kw │     4.50w │     4.50w │     16.69% │
│ [bench/micro/fiber_bench.ml] ivar                │  28.96us │ 49.08kw │    15.54w │    15.54w │     64.41% │
│ [bench/micro/fiber_bench.ml] Var.set             │  44.96us │ 42.08kw │ 1_454.95w │ 1_454.95w │    100.00% │
│ [bench/micro/fiber_bench.ml] Var.get             │  24.28us │ 32.12kw │     9.58w │     9.58w │     54.01% │
│ [bench/micro/fiber_bench.ml] catching exceptions │  24.75us │ 20.14kw │   210.36w │   210.36w │     55.05% │
│ [bench/micro/fiber_bench.ml] installing handlers │  34.07us │ 77.07kw │    19.64w │    19.64w │     75.78% │
└──────────────────────────────────────────────────┴──────────┴─────────┴───────────┴───────────┴────────────┘
```